% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/departures.R
\name{classify_departure_status}
\alias{classify_departure_status}
\title{Classify CUMTD Bus Departures Status}
\usage{
classify_departure_status(departures_info)
}
\arguments{
\item{departures_info}{Results returned from \code{\link[cumtd:get_departures_by_stop]{cumtd::get_departures_by_stop()}}.}
}
\value{
An augmented \code{data.frame} that includes:
\itemize{
\item \code{late}          is the bus late
\item \code{delay}         any delay?
}
}
\description{
Provides quick computations for
}
\examples{
\dontrun{
# Retrieve departure information from API
departures = get_departures_by_stop("iu")

# Add columns related to late and delay amount.
classified_departures = classify_departure_status(departures)
}
}
