% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-backend.R
\name{convert_response}
\alias{convert_response}
\title{Convert Response to R Object}
\usage{
convert_response(response, flatten = TRUE)
}
\arguments{
\item{response}{The result from the API query.}

\item{flatten}{Ensure that only one `data.frame`` is returned.}
}
\description{
Converts the response from JSON to an R \code{list}/\code{data.frame}.
}
