% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumtd-package.R
\docType{package}
\name{cumtd-package}
\alias{cumtd-package}
\title{cumtd: R interface to Champaign–Urbana Mass Transit District Web API}
\description{
Provides access to the realtime Web API run by Champaign–Urbana Mass Transit District ('CUMTD').
}
\details{
Access data from CUMTD's API
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rd.thecoatlessprofessor.com/r-web-api/}
}

}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{balamut2@illinois.edu} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Champaign-Urbana Mass Transit District (CUMTD) \email{developer@cumtd.com} [copyright holder]
}

}
