% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-backend.R
\name{cumtd}
\alias{cumtd}
\title{Construct and Call a "GET" Request to CUMTD API}
\usage{
cumtd(endpoint, ..., key = NULL, version = "v2.2", format = c("json",
  "xml"))
}
\arguments{
\item{endpoint}{Resource path}

\item{...}{Additional parameters}

\item{key}{Authorization}

\item{version}{API version to use}

\item{format}{Response type}
}
\value{
An \code{httr} response object.
}
\description{
Builds the "GET" request based on the resource being requested. Once built,
the call is initiated.
}
\examples{
\dontrun{
# Requires a key to retrieve data
cumtd("getapiusage")
}
}
