% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-keys.R
\name{cumtd_show_key}
\alias{cumtd_show_key}
\alias{cumtd_check_key}
\alias{cumtd_set_key}
\title{Sets, Retrieve, and Check the CUMTD API Key}
\usage{
cumtd_show_key()

cumtd_check_key(key = NULL)

cumtd_set_key(key)
}
\arguments{
\item{key}{The CUMTD API assigned}
}
\description{
Suite of functions for handling the \code{CUMTD_KEY} used in API calls.
}
\details{
Obtain a CUMTD API key from \url{https://developer.cumtd.com/}
}
\examples{
cumtd_show_key()
\dontrun{
cumtd_check_key()
}
\dontrun{
cumtd_set_key("your-key-here")
}
}
