% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routes.R
\name{get_all_routes}
\alias{get_all_routes}
\title{Retrieve CUMTD Routes}
\usage{
get_all_routes()
}
\value{
A \code{data.frame} containing the route information with columns:
\itemize{
\item \code{route_color}	    hex color of route
\item \code{route_id}	        id of route
\item \code{route_long_name}	long name
\item \code{route_short_name}	short name
\item \code{route_text_color}	hex color of route text
}
}
\description{
Obtains a list of CUMTD routes
}
\references{
\url{https://developer.cumtd.com/documentation/v2.2/method/getroutes/}
}
