% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops.R
\name{get_all_stops}
\alias{get_all_stops}
\title{Retrieve CUMTD Bus Stop Locations}
\usage{
get_all_stops()
}
\value{
A \code{data.frame} containing the route information with columns:
\itemize{
\item \code{code}	      text message code
\item \code{stop_point} stop points that compose a parent stop
\item \code{stop_id}	  id of stop
\item \code{stop_lat}	  latitude of stop
\item \code{stop_lon}	  longitude of stop
\item \code{stop_name}	name of stop
}
}
\description{
Obtains a list of CUMTD bus stop locations
}
\references{
\url{https://developer.cumtd.com/documentation/v2.2/method/getstops/}
}
