% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vehicles.R
\name{get_all_vehicles}
\alias{get_all_vehicles}
\title{Retrieve All CUMTD Vehicles that are Tracked}
\usage{
get_all_vehicles()
}
\value{
A \code{data.frame} containing the vehicle information with columns:
\itemize{
\item \code{vehicle_id	}        Vehicle number associated with vehicle
\item \code{trip}                Trip information for the departure
\item \code{location}	           The last known latitude and longitude of the vehicle
\item \code{previous_stop_id}	   The last stop that the vehicle served
\item \code{next_stop_id}	       The next stop that the vehicle will serve
\item \code{origin_stop_id}	     The stop where the vehicle began it's trip
\item \code{destination_stop_id} The stop where the vehicle will end its trip
\item \code{last_updated}	       The last time the vehicle sent a real-time location
update to our system.
}
}
\description{
Get information for all currently tracked vehicles.
}
\references{
\url{https://developer.cumtd.com/documentation/v2.2/method/getvehicles/}
}
