% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage.R
\name{get_api_usage}
\alias{get_api_usage}
\title{Retrieve API Usage for specific methods}
\usage{
get_api_usage(
  method = NULL,
  version = NULL,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{method}{(optional) a single API method for which you want data
(e.g. GetDeparturesByStop)}

\item{version}{(optional) a single API version for which you want data
(e.g. 2.1)}

\item{start_date}{(optional) the first day for which you want data, in
YYYY-MM-DD format (cannot be earlier than 2013-02-07)}

\item{end_date}{(optional) the last day for which you want data, in
YYYY-MM-DD format (cannot be later than yesterday)}
}
\value{
\itemize{
\item \code{date}	    a date in YYYY-MM-DD format
\item \code{calls}	  the number of requests made to a method
\item \code{versions}	a list of versions of the API that were used on a given date
\item \code{number}	  an API version number
\item \code{methods}	a list of API methods that were used on a given date and within a given API version
\item \code{name}   	the name of an API method
}
}
\description{
Get a summary of the API usage for your key.
}
