% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/departures.R
\name{get_departures_by_stop}
\alias{get_departures_by_stop}
\title{Retrieve a List of CUMTD Bus Departures}
\usage{
get_departures_by_stop(stop_id, route_id = NULL, pt = NULL,
  count = NULL)
}
\arguments{
\item{stop_id}{(required) id of the stop (ex. IT:1 or IT)}

\item{route_id}{(optional) vector of route ids}

\item{pt}{(optional) preview time in minutes between 0 and 60 (30 by default)}

\item{count}{(optional) maximum number of departures you would like to recieve}
}
\value{
A \code{data.frame} with departure information for a specific stop:
\itemize{
\item \code{destination}	 trip's destination stop
\item \code{expected}	     expected departure time of the bus for the given stop
\item \code{expected_mins} number of minutes before expected departure time
\item \code{headsign}	     information usually shown on headsign
\item \code{location}	     latitude and longitude of vehicle
\item \code{is_monitored}	 whether the vehicle is communicating
\item \code{is_scheduled}	 whether this trip was scheduled
\item \code{is_istop}	     if this trip can be boarded without a fare/pass
\item \code{origin}	       trip's origin stop
\item \code{route}     	   route information for the trip
\item \code{scheduled}	   scheduled departure time of the bus for the given stop
\item \code{stop_id}	     id of the stop the bus will be at
\item \code{trip}	         trip information for the departure
\item \code{vehicle_id}	   id associated with vehicle
}
}
\description{
Get a list of real-time departures for a specific \code{stop_id}.
}
