% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{get_route_shape}
\alias{get_route_shape}
\title{Retrieve CUMTD Route Shapes}
\usage{
get_route_shape(shape_id)
}
\arguments{
\item{shape_id}{(required) id of the shape}
}
\value{
A \code{data.frame} containing the route information with columns:
\itemize{
\item \code{shape_dist_traveled} total distance traveled to this point
\item \code{shape_pt_lat}        latitude of point
\item \code{shape_pt_lon}	       longitude of point
\item \code{stop_id}	           the stop id associeated with the shape point.
\item \code{shape_pt_sequence}	 sequence of point in GTFS feed
}
}
\description{
Get a list of points that define the shape of the route on a map.
}
\references{
\url{https://developer.cumtd.com/documentation/v2.2/method/getshape/}
}
