% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{get_route_shape_between_stops}
\alias{get_route_shape_between_stops}
\title{Retrieve CUMTD Route Shapes Between Stops}
\usage{
get_route_shape_between_stops(begin_stop_id, end_stop_id, shape_id)
}
\arguments{
\item{begin_stop_id}{(required) id of the beginning stop point}

\item{end_stop_id}{(required) id of the ending stop point}

\item{shape_id}{(required) id of the shape}
}
\value{
A \code{data.frame} containing the route information with columns:
\itemize{
\item \code{shape_dist_traveled} total distance traveled to this point
\item \code{shape_pt_lat}        latitude of point
\item \code{shape_pt_lon}	       longitude of point
\item \code{stop_id}	           the stop id associeated with the shape point.
\item \code{shape_pt_sequence}	 sequence of point in GTFS feed
}
}
\description{
Get a list of points that define part of a shape on a map between two stops.
}
\references{
\url{https://developer.cumtd.com/documentation/v2.2/method/getshape/}
}
