% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumtd-package.R
\docType{package}
\name{cumtd-package}
\alias{cumtd-package}
\alias{_PACKAGE}
\title{cumtd: R interface to Champaign–Urbana Mass Transit District Web API}
\description{
Provides access to the realtime Web API run by
    Champaign–Urbana Mass Transit District ('CUMTD').
}
\details{
Access data from CUMTD's API
}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{balamut2@illinois.edu} (0000-0003-2826-8458) [copyright holder]

Other contributors:
\itemize{
  \item Champaign-Urbana Mass Transit District (CUMTD) \email{developer@cumtd.com} [copyright holder]
}

}
